separate DF SDK classes from DF worker classes
fix typo

remove TODOs

remove commented-out code
