+++ src/DurableSDK/Actions/ExternalEventAction.cs
