+++ src/DurableSDK/ActivityFailureException.cs
