+++ src/DurableSDK/Commands/SetFunctionInvocationContextCommand.cs
