+++ src/DurableSDK/DurableActivityErrorHandler.cs
