+++ src/DurableSDK/HistoryEvent.cs
