+++ src/DurableSDK/Tasks/ActivityInvocationTask.cs
@@ -16,6 +16,7 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Durable.Tasks
     using Microsoft.Azure.Functions.PowerShellWorker;
     using Microsoft.Azure.Functions.PowerShellWorker.Durable;
     using Microsoft.Azure.Functions.PowerShellWorker.Durable.Actions;
+    using Microsoft.Azure.Functions.PowerShellWorker.DurableWorker;
 
     public class ActivityInvocationTask : DurableTask
     {
