+++ src/DurableSDK/Tasks/ExternalEventTask.cs
